Control jsonpatches generate to match Tomviz semantics
When dealing with Tomviz state the id associated with a object
should be treated as immutable. However, jsonpatch as it stands
will try to perform replaces on ids, which make not sense for
Tomviz pipelines. This small patch ensures that a remove and add
patch sequence is used instead. Not sure how we could make this
upstreamable, so for now maintaining a fork.